Prevent concurrent calls to droid_media_queue and droid_media_(drain|stop).
The decoder stream is unlocked while queuing frames and when waiting for
an EOS signal. In the right circumstances that can lead the finish
call advancing to calling stop during a call to queue which can in turn
mistakenly restart the codec.  Locking while this happens leads to
deadlocks beecause of feedback from interdependent callbacks, but we can
emulate an asymetric lock by setting a boolean protected by the state
mutex while the queue is executing and waiting for a signal in the
finish call if the boolean is set.